test: score gene essentiality per task category via /run gene-essentiality
Replaces the all-task binary scoring with the graded, task-scoped analysis, so a
gene essential only for a capability task (SU, BS, IC) is no longer counted as a
proliferation prediction alongside viability tasks (GR, ER). See issue #1076.

- gene-essentiality.yml runs gradedEssentiality.py, which now overwrites the usual
  gene-essential.csv and gene-essential_summary.md; the posted comment explains the
  viability/capability split.
- geneEssentiality.py is removed: the graded run reports the all-task scoring too,
  so nothing is lost.
- The graded scoring (evaluate_graded, summary_to_markdown, auroc, auprc) lives in
  evaluateHart2015Essentiality.py, which needs only pandas, so /reevaluate
  gene-essentiality keeps working from the committed matrix without a solve.
- reevaluateGeneEssentiality.py parses the new matrix, which carries the task
  categories and growth ratio per gene.
- The growth ratio is rounded to a fixed precision before scoring and average
  precision is computed over tied scores as a group, so the summary regenerated
  from the committed matrix is identical to the one the run produced.
- Results regenerated from the graded run; see the pull request for provenance.
