+++ data/testResults/gene-essential_summary.md
@@ -1,10 +1,10 @@
-### Gene essentiality vs Hart 2015 fitness genes
+### Graded gene essentiality vs Hart 2015 (task-scope analysis)
 
-| cellLine | TP | TN | FP | FN | accuracy | sensitivity | specificity | F1 | MCC |
-| --- | --- | --- | --- | --- | --- | --- | --- | --- | --- |
-| DLD1 | 125 | 2172 | 124 | 224 | 0.8684 | 0.3582 | 0.946 | 0.4181 | 0.3525 |
-| GBM | 111 | 2145 | 138 | 251 | 0.8529 | 0.3066 | 0.9396 | 0.3633 | 0.2897 |
-| HCT116 | 141 | 2189 | 130 | 246 | 0.861 | 0.3643 | 0.9439 | 0.4286 | 0.3595 |
-| HELA | 114 | 2233 | 164 | 197 | 0.8667 | 0.3666 | 0.9316 | 0.3871 | 0.3132 |
-| RPE1 | 86 | 2183 | 162 | 214 | 0.8578 | 0.2867 | 0.9309 | 0.3139 | 0.2367 |
-| all | 59 | 2355 | 159 | 69 | 0.9137 | 0.4609 | 0.9368 | 0.341 | 0.3103 |
+| cellLine | allTaskMCC | allTaskFP | viabilityMCC | viabilityFP | growthAUROC | growthAUPRC | baseRate | capabilityOnly |
+| --- | --- | --- | --- | --- | --- | --- | --- | --- |
+| DLD1 | 0.3754 | 126 | 0.3859 | 93 | 0.6766 | 0.3368 | 0.1684 | 49 |
+| GBM | 0.2985 | 160 | 0.3308 | 119 | 0.658 | 0.2938 | 0.1624 | 50 |
+| HCT116 | 0.3946 | 124 | 0.4049 | 86 | 0.6839 | 0.3673 | 0.1812 | 56 |
+| HELA | 0.3253 | 154 | 0.3337 | 117 | 0.6709 | 0.2937 | 0.1467 | 49 |
+| RPE1 | 0.2405 | 191 | 0.2807 | 144 | 0.6478 | 0.2401 | 0.136 | 52 |
+| all |  |  |  |  | 0.6678 | 0.3042 | 0.1592 |  |
